refactor: Board 컴포넌트의 useBoard와 useTouchTracking 커스텀 훅을 별도 파일로 분리
